* Реализован базовый класс запросов к базе данных, который содержит реализованные методы для типичных запросов к таблицам, такие как запрсо на проверку существованяи таблицы, запрос на получение количества строк в таблице. Каждый класс, который описывает является обёрткой над таблицей наследуется от данного класса. * В класс Dictionary в методы, работающие с таблицей добавлена проверка на то, что таблица существует.

+++ app/src/main/java/knack/college/learnenglish/model/Dictionary.java
@@ -4,19 +4,21 @@ import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
+import android.widget.Toast;
 
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 import java.util.UUID;
 
+import knack.college.learnenglish.R;
 import knack.college.learnenglish.exceptions.EmptyData;
 import knack.college.learnenglish.exceptions.MoreMaxSymbols;
 import knack.college.learnenglish.exceptions.NoEnglishWord;
 import knack.college.learnenglish.exceptions.NoRussianWord;
+import knack.college.learnenglish.model.database.BaseQueries;
 import knack.college.learnenglish.model.database.DictionaryContract;
 import knack.college.learnenglish.model.database.LearnEnglishDatabaseHelper;
+import knack.college.learnenglish.model.toasts.ToastWrapper;
 
 import static knack.college.learnenglish.model.database.DictionaryContract.Dictionary.DICTIONARY_TABLE_NAME;
 import static knack.college.learnenglish.model.database.DictionaryContract.Dictionary.ENGLISH_WORD_COLUMN_NAME;
@@ -29,7 +31,7 @@ import static knack.college.learnenglish.model.database.DictionaryContract.Dicti
 import static knack.college.learnenglish.model.database.DictionaryContract.Dictionary.getDropDictionaryTableQuery;
 
 /** Класс, описывающий словарь */
-public class Dictionary {
+public class Dictionary extends BaseQueries {
 
     private static final String WORD_MORE_MAX_SYMBOLS_EXCEPTION_MESSAGE =
             "Слово, которое вы ввели больше 255 символов.";
@@ -40,25 +42,66 @@ public class Dictionary {
     private static final String NO_DATA_EXCEPTION_MESSAGE = "Не хватает данных для заполнения";
 
     private Context context;
-    private SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-    private String currentDate = sdf.format(new Date());
 
+    private ToastWrapper toast;
     private Validator validator;
     private LearnEnglishDatabaseHelper helper;
     private SQLiteDatabase database;
 
-    public Dictionary(Context c) {
+    public Dictionary(Context c) throws Exception {
+        super(c);
+
         context = c;
 
-        validator = new Validator();
-        helper = new LearnEnglishDatabaseHelper(context);
-        database = helper.getWritableDatabase();
+        initializeToast();
+        initializeValidator();
+        initializeDatabaseHelper();
+        initializeDatabase();
     }
 
-    public Dictionary() {
-        validator = new Validator();
-        helper = new LearnEnglishDatabaseHelper(context);
-        database = helper.getWritableDatabase();
+    private void initializeToast() {
+        try {
+            toast = new ToastWrapper(context);
+        } catch (Exception e) {
+            Toast.makeText(
+                    context,
+                    context.getResources().getString(R.string.error_message_failed_initialize_toast),
+                    Toast.LENGTH_LONG
+            ).show();
+        }
+    }
+
+    private void initializeValidator() {
+        try {
+            validator = new Validator();
+        } catch (Exception e) {
+            toast.show(
+                    context.getResources().getString(R.string
+                            .error_message_failed_initialize_validator)
+            );
+        }
+    }
+
+    private void initializeDatabaseHelper() {
+        try {
+            helper = new LearnEnglishDatabaseHelper(context);
+        } catch (Exception e) {
+            toast.show(
+                    context.getResources().getString(R.string
+                            .error_message_failed_initialize_database_helper)
+            );
+        }
+    }
+
+    private void initializeDatabase() {
+        try {
+            database = helper.getWritableDatabase();
+        } catch (Exception e) {
+            toast.show(
+                    context.getResources().getString(R.string
+                            .error_message_failed_initialize_database)
+            );
+        }
     }
 
     /** Метод, который добавляет слово на иностранном языке и слово-перевод в словарь */
@@ -69,85 +112,97 @@ public class Dictionary {
                         && !validator.isWordMoreMaxSymbols(translate)) {
                     if (validator.isEnglishCharactersInWord(englishWord)) {
                         if (validator.isRussianCharactersInWord(translate)) {
-                            ContentValues values = new ContentValues();
+                            if (isExistTable(null, DictionaryContract.Dictionary
+                                    .DICTIONARY_TABLE_NAME)) {
+                                ContentValues values = new ContentValues();
 
-                            values.put(GUID_COLUMN_NAME, UUID.randomUUID().toString());
-                            values.put(ENGLISH_WORD_COLUMN_NAME, englishWord);
-                            values.put(TRANSLATE_WORD_COLUMN_NAME, translate);
+                                values.put(GUID_COLUMN_NAME, UUID.randomUUID().toString());
+                                values.put(ENGLISH_WORD_COLUMN_NAME, englishWord);
+                                values.put(TRANSLATE_WORD_COLUMN_NAME, translate);
 
-                            database.insert(DICTIONARY_TABLE_NAME, null, values);
-                        } else {
-                            throw new NoRussianWord(NO_RUSSIAN_WORD_EXCEPTION_MESSAGE);
-                        }
-                    } else {
-                        throw new NoEnglishWord(NO_ENGLISH_WORD_EXCEPTION_MESSAGE);
-                    }
-                } else {
-                    throw new MoreMaxSymbols(WORD_MORE_MAX_SYMBOLS_EXCEPTION_MESSAGE);
-                }
-            } else {
-                throw new EmptyData(NO_DATA_EXCEPTION_MESSAGE);
-            }
+                                database.insert(DICTIONARY_TABLE_NAME, null, values);
+                            } else {
+                                helper.onCreate(database);
+
+                                ContentValues values = new ContentValues();
+
+                                values.put(GUID_COLUMN_NAME, UUID.randomUUID().toString());
+                                values.put(ENGLISH_WORD_COLUMN_NAME, englishWord);
+                                values.put(TRANSLATE_WORD_COLUMN_NAME, translate);
+
+                                database.insert(DICTIONARY_TABLE_NAME, null, values);
+                            }
+                        } else throw new NoRussianWord(NO_RUSSIAN_WORD_EXCEPTION_MESSAGE);
+                    } else throw new NoEnglishWord(NO_ENGLISH_WORD_EXCEPTION_MESSAGE);
+                } else throw new MoreMaxSymbols(WORD_MORE_MAX_SYMBOLS_EXCEPTION_MESSAGE);
+            } else throw new EmptyData(NO_DATA_EXCEPTION_MESSAGE);
         }
     }
 
     /** Метод, который удаляет словарь */
     public void delete() throws Exception {
-        database.execSQL(getDropDictionaryTableQuery().toString());
+        if (isExistTable(null, DictionaryContract.Dictionary.DICTIONARY_TABLE_NAME))
+            database.execSQL(getDropDictionaryTableQuery().toString());
     }
 
     /** Метод, который очищяет словарь*/
     public void clear() throws Exception {
-        database.execSQL(getDeleteAllRowsInTableQuery().toString());
+        if (isExistTable(null, DictionaryContract.Dictionary.DICTIONARY_TABLE_NAME))
+            database.execSQL(getDeleteAllRowsInTableQuery().toString());
     }
 
     /** Метод, который курсор со всеми строками и столбцами из словаря */
     private Cursor getAllWordsCursor() throws Exception {
-        String[] columnSelectList = {
-                OUID_COLUMN_NAME,
-                GUID_COLUMN_NAME,
-                ENGLISH_WORD_COLUMN_NAME,
-                TRANSLATE_WORD_COLUMN_NAME,
-                LAST_TRAINING_DATE
-        };
+        if (isExistTable(null, DictionaryContract.Dictionary.DICTIONARY_TABLE_NAME)) {
+            String[] columnSelectList = {
+                    OUID_COLUMN_NAME,
+                    GUID_COLUMN_NAME,
+                    ENGLISH_WORD_COLUMN_NAME,
+                    TRANSLATE_WORD_COLUMN_NAME,
+                    LAST_TRAINING_DATE
+            };
 
-        return database.query(
-                DictionaryContract.Dictionary.DICTIONARY_TABLE_NAME,
-                columnSelectList,
-                null,
-                null,
-                null,
-                null,
-                null
-        );
+            return database.query(
+                    DictionaryContract.Dictionary.DICTIONARY_TABLE_NAME,
+                    columnSelectList,
+                    null,
+                    null,
+                    null,
+                    null,
+                    null
+            );
+        } else return null;
     }
 
     /** Метод, который возвращает список слов со словаря */
-    public List<WordFromDictionary> getAllWordsList() throws Exception {
+    public List getAllWordsList() throws Exception {
         List<WordFromDictionary> allWordsList = new ArrayList<>();
 
         Cursor cursor = null;
 
         try {
-            cursor = getAllWordsCursor();
+            if (isExistTable(null, DictionaryContract.Dictionary.DICTIONARY_TABLE_NAME)) {
+                cursor = getAllWordsCursor();
 
-            while (cursor.moveToNext()) {
-                WordFromDictionary word = new WordFromDictionary();
-                word.setOuid(cursor.getString(cursor.getColumnIndex(OUID_COLUMN_NAME)));
-                word.setGuid(cursor.getString(cursor.getColumnIndex(GUID_COLUMN_NAME)));
-                word.setEnglishWord(
-                        cursor.getString(cursor.getColumnIndex(ENGLISH_WORD_COLUMN_NAME))
-                );
-                word.setTranslateWord(
-                        cursor.getString(cursor.getColumnIndex(TRANSLATE_WORD_COLUMN_NAME))
-                );
+                if (cursor != null) {
+                    while (cursor.moveToNext()) {
+                        WordFromDictionary word = new WordFromDictionary();
+                        word.setOuid(cursor.getString(cursor.getColumnIndex(OUID_COLUMN_NAME)));
+                        word.setGuid(cursor.getString(cursor.getColumnIndex(GUID_COLUMN_NAME)));
+                        word.setEnglishWord(
+                                cursor.getString(cursor.getColumnIndex(ENGLISH_WORD_COLUMN_NAME))
+                        );
+                        word.setTranslateWord(
+                                cursor.getString(cursor.getColumnIndex(TRANSLATE_WORD_COLUMN_NAME))
+                        );
 
-                allWordsList.add(word);
+                        allWordsList.add(word);
+                    }
+                }
             }
         } finally {
-            if (cursor != null) {
+            if (cursor != null)
                 cursor.close();
-            }
         }
 
         return allWordsList;
@@ -155,11 +210,20 @@ public class Dictionary {
 
     /** Метод, который возвращает количество слов в словаре */
     public int getNumberOfWords() throws Exception {
-        int count;
-        Cursor cursor = database.rawQuery(getCountRowsInTableQuery().toString(), null);
-        cursor.moveToFirst();
-        count = cursor.getInt(0);
-        cursor.close();
+        int count = 0;
+        Cursor cursor = null;
+        try {
+            if (isExistTable(null, DictionaryContract.Dictionary.DICTIONARY_TABLE_NAME)) {
+                cursor = database.rawQuery(getCountRowsInTableQuery().toString(), null);
+                cursor.moveToFirst();
+                count = cursor.getInt(0);
+            }
+        } catch (Exception e) {
+            count = 0;
+        } finally {
+            if (cursor != null)
+                cursor.close();
+        }
 
         return count;
     }
@@ -167,81 +231,14 @@ public class Dictionary {
     /** Метод, который удаляет слово из словаря */
     public void deleteWord(String wordGuid) throws Exception {
         if (wordGuid != null && !wordGuid.isEmpty()) {
-            LearnEnglishDatabaseHelper helper = new LearnEnglishDatabaseHelper(context);
-            SQLiteDatabase database = helper.getWritableDatabase();
-            database.delete(
-                    DICTIONARY_TABLE_NAME,
-                    GUID_COLUMN_NAME + " = " + "'" + wordGuid + "'",
-                    null
-            );
-        }
-    }
-
-    /** Метод, который слову сетает последнюю дату тренировки */
-    public void setLastTrainingWordDate(String wordGuid) throws Exception {
-        if (wordGuid != null && !wordGuid.isEmpty()) {
-            LearnEnglishDatabaseHelper helper = new LearnEnglishDatabaseHelper(context);
-            SQLiteDatabase db = helper.getReadableDatabase();
-
-            // Current date
-            SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-            String strDate = sdf.format(new Date());
-
-            // New value
-            ContentValues values = new ContentValues();
-            values.put(LAST_TRAINING_DATE, strDate);
-
-            db.update(
-                    DICTIONARY_TABLE_NAME,
-                    values,
-                    GUID_COLUMN_NAME + " = " + "'" + wordGuid + "'",
-                    null
-            );
-        }
-    }
-
-    /** Метод, который возвращает курсор с словами, в которых не проставлена
-     * дата последней тренировки
-     * или с момента тренировки прошло больше 1 дня */
-    public Cursor getForgottenWordsCursor() throws Exception {
-        return database.rawQuery("SELECT " + OUID_COLUMN_NAME + ", " + GUID_COLUMN_NAME + ", "
-                + ENGLISH_WORD_COLUMN_NAME + ", " + TRANSLATE_WORD_COLUMN_NAME + ", "
-                + LAST_TRAINING_DATE + " FROM DICTIONARY AS dictionary WHERE julianday('"
-                + currentDate
-                + "') - julianday(dictionary.LAST_TRAINING_DATE) >= 1 OR "
-                + "dictionary.LAST_TRAINING_DATE IS NULL" , null);
-    }
-
-    /** Метод, который возвращает коллекцию слов, в которых не проставлена
-     * дата последней тренировки
-     * или с момента тренировки прошло больше 1 дня */
-    public List<WordFromDictionary> getForgottenWords() throws Exception {
-        List<WordFromDictionary> allWordsList = new ArrayList<>();
-        Cursor cursor = null;
-
-        try {
-            cursor = getForgottenWordsCursor();
-
-            while (cursor.moveToNext()) {
-                WordFromDictionary word = new WordFromDictionary();
-                word.setOuid(cursor.getString(cursor.getColumnIndex(OUID_COLUMN_NAME)));
-                word.setGuid(cursor.getString(cursor.getColumnIndex(GUID_COLUMN_NAME)));
-                word.setEnglishWord(
-                        cursor.getString(cursor.getColumnIndex(ENGLISH_WORD_COLUMN_NAME))
+            if (isExistTable(null, DictionaryContract.Dictionary.DICTIONARY_TABLE_NAME)) {
+                database.delete(
+                        DICTIONARY_TABLE_NAME,
+                        GUID_COLUMN_NAME + " = " + "'" + wordGuid + "'",
+                        null
                 );
-                word.setTranslateWord(
-                        cursor.getString(cursor.getColumnIndex(TRANSLATE_WORD_COLUMN_NAME))
-                );
-
-                allWordsList.add(word);
-            }
-        } finally {
-            if (cursor != null) {
-                cursor.close();
             }
         }
-
-        return allWordsList;
     }
 
     /** Метод, который восстанавливает слово в словаре */
